Add some options for configuring transport to upstream server
We use this in local dev, but many tools can't handle s2 compression
or native Protobuf format, which makes integration a little tricky.

It would also be great to be able to point this to a backend on
something like `http://localhost:8080`, but HTTPS is baked into the
the upstream `psdb` library, so that would need updating first:
https://github.com/planetscale/psdb/blob/5e0969e17235f6a9229f5c1ab7208588c514b421/core/client/client.go#L76

Naming things is hard, but the option names and abbreviations have
been chosen with the `mysql` CLI in mind.